backend/requirements.txt
flask==3.1.1
flask-cors==5.0.1
bcrypt==4.3.0
pyjwt==2.10.1
requests==2.32.3
cryptography==43.0.3
gunicorn==23.0.0
